# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  name: string;
  ageOrClassLevel: string;
  currentStreak: number;
  longestStreak: number;
}

export interface SignupInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  ageOrClassLevel: string;
  /** @pattern ^[0-9]{4,6}$ */
  pin: string;
}

export interface LoginInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @pattern ^[0-9]{4,6}$ */
  pin: string;
}

export interface AuthSession {
  student: Student;
}

export type MessageSender = typeof MessageSender[keyof typeof MessageSender];


export const MessageSender = {
  student: 'student',
  temmy: 'temmy',
} as const;

export type MessageContentType = typeof MessageContentType[keyof typeof MessageContentType];


export const MessageContentType = {
  text: 'text',
  voice: 'voice',
  image: 'image',
} as const;

export interface Message {
  id: number;
  sender: MessageSender;
  contentType: MessageContentType;
  contentText: string;
  createdAt: string;
}

export type ChatMessageInputContentType = typeof ChatMessageInputContentType[keyof typeof ChatMessageInputContentType];


export const ChatMessageInputContentType = {
  text: 'text',
  voice: 'voice',
  image: 'image',
} as const;

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  contentText: string;
  contentType?: ChatMessageInputContentType;
  /** @maxLength 8000000 */
  mediaData?: string;
  /** @maxLength 100 */
  mediaMimeType?: string;
}

export interface ProgressTopic {
  subject: string;
  topic: string;
  timesDiscussed: number;
  lastCoveredAt?: string;
}

export interface Badge {
  badgeType: string;
  awardedAt: string;
}

export interface Progress {
  currentStreak: number;
  longestStreak: number;
  topics: ProgressTopic[];
  badges: Badge[];
}

export interface ChatReply {
  studentMessage: Message;
  temmyMessage: Message;
  progress: Progress;
}