# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageInputContentType } from './chatMessageInputContentType';

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  contentText: string;
  contentType?: ChatMessageInputContentType;
  /** @maxLength 8000000 */
  mediaData?: string;
  /** @maxLength 100 */
  mediaMimeType?: string;
}