# lib/api-zod/src/generated/types/chatReply.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';
import type { Progress } from './progress';

export interface ChatReply {
  studentMessage: Message;
  temmyMessage: Message;
  progress: Progress;
}