# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @pattern ^[0-9]{4,6}$ */
  pin: string;
}