# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageContentType } from './messageContentType';
import type { MessageSender } from './messageSender';

export interface Message {
  id: number;
  sender: MessageSender;
  contentType: MessageContentType;
  contentText: string;
  createdAt: Date;
}