# lib/api-zod/src/generated/types/progress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';
import type { ProgressTopic } from './progressTopic';

export interface Progress {
  currentStreak: number;
  longestStreak: number;
  topics: ProgressTopic[];
  badges: Badge[];
}